Updated double-check to make sure path doesn't already exist
git-svn-id: svn+ssh://lyra.berkeley.edu/Bloom/amorgan/SVNREPO/q_soft@228 15f48517-4a6d-0410-94cc-8a7d56af1ecd

+++ trunk/Software/AutoRedux/q_insert.py
@@ -34,7 +34,10 @@ def grab_script(filepath,rm=True):
             print 'Unrecognized line in script!'
             print line
     if rm:
-        os.system('mv ' + filepath + ' ' + finished_scripts_path + '.')
+        if not os.path.exists(finished_scripts_path):
+            os.system('mv ' + filepath + ' ' + finished_scripts_path + '.')
+        else:
+            print "**FINISHED SCRIPTS PATH ALREADY EXISTS.  Not moving."
         
 
 def queue_insert(reduction_command):
